마니또 배정 이전에도 구독 가능하도록 버튼 만들었다.
