+++ src/Connections/NovaConnection.js
@@ -2,7 +2,7 @@ import axios from "axios";
 import { getToken } from "../Utils/Cookie";
 
 const instance = axios.create({
-  baseURL: "https://api.nova.snuaaa.net:9887/manitto",
+  baseURL: "https://api.nova.snuaaa.net:9885/manitto",
 });
 
 const request = {
